requirements.txt
flask-restplus==0.9.2
gphoto2==1.8.3
RPi.GPIO==0.6.3
pillow==5.2.0
